Move Authn Status to own controller
Some additional bootstrapping was required to allow for this (in
services.yaml).
